# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Users CRUD, Login & Logout API For Testing",
        "description": "API CRUD de usuarios, Login & Logout para Pruebas ",
        "version": "0.1b"
    },
    "servers": [
        {
            "url": "http://localhost/api-auth-tablar/public/api"
        }
    ],
    "paths": {
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "AUTH"
                ],
                "summary": "Login del Usuario",
                "operationId": "login",
                "requestBody": {
                    "description": "Datos de Autenticación para Inicio de Sesión",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "usernameOrEmail",
                                    "password"
                                ],
                                "properties": {
                                    "usernameOrEmail": {
                                        "description": "Nombre de Usuario o Correo Electrónico",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitosa"
                    },
                    "401": {
                        "description": "Usuario no autorizado"
                    },
                    "400": {
                        "description": "Credenciales Incorrecta"
                    }
                }
            }
        },
        "/v1/auth/logout": {
            "get": {
                "tags": [
                    "AUTH"
                ],
                "summary": "Logout del Usuario",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitosa"
                    }
                }
            }
        },
        "/v1/users": {
            "get": {
                "tags": [
                    "USERS"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Mostrar listado de Usuarios",
                "operationId": "9b1f09ab22ffe6f8c099f1614bd03deb",
                "responses": {
                    "200": {
                        "description": "Lista de todos los Usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "fullname": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "birthday": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1981-04-01"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "3351425869"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Urb. Virgen del Carmen, Conjunto Residencial La Floresta"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "Venezuela"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Puerto La Cruz"
                                                    },
                                                    "email": {
                                                        "type": "email",
                                                        "example": "john.doe43@example.com"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "jdoe43"
                                                    },
                                                    "status": {
                                                        "type": "boolean",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Lista Vacía",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}